+++ helpers.c
@@ -10,8 +10,8 @@
 
 
 // Calculates the distance of p from a reference point, given in the form of an array.
-double calculateDistanceArray(double *p, int start, double *ref, uint dims) {
-	double distance = 0;
+float calculateDistanceArray(float *p, int start, float *ref, uint dims) {
+	float distance = 0;
 	for (int i = 0; i < dims; i++) {
 		distance += pow(p[start + i] - ref[i], 2);
 	}
@@ -22,7 +22,7 @@ double calculateDistanceArray(double *p, int start, double *ref, uint dims) {
 
 // Standard Lomuto partition function.
 // Taken from https://www.geeksforgeeks.org/quickselect-a-simple-iterative-implementation/
-uint partition(double *arr, uint low, uint high) {
+uint partition(float *arr, uint low, uint high) {
 	uint pivot = arr[high];
 
 	uint i = low - 1;
@@ -31,13 +31,13 @@ uint partition(double *arr, uint low, uint high) {
 			i++;
 
 			// Swap arr[i] and arr[j].
-			double tmp = arr[i];
+			float tmp = arr[i];
 			arr[i] = arr[j];
 			arr[j] = tmp;
 		}
 	}
 
-	double tmp = arr[i + 1];
+	float tmp = arr[i + 1];
 	arr[i + 1] = arr[high];
 	arr[high] = tmp;
 
@@ -48,7 +48,7 @@ uint partition(double *arr, uint low, uint high) {
 // Finds the k-th smallest element of an array. Kept outside of quickselect,
 // since the distances array will probably consist of an even number of floats,
 // which means we need to find the semi-sum of the values in the middle.
-double kthSmallest(double *array, uint start, uint end, uint k) {
+float kthSmallest(float *array, uint start, uint end, uint k) {
 	while (start <= end) {
 		uint pivotIndex = partition(array, start, end);
 
@@ -68,18 +68,18 @@ double kthSmallest(double *array, uint start, uint end, uint k) {
 
 
 // Finds the median in an array of distances.
-double quickselect(double *distances, uint end) {
+float quickselect(float *distances, uint end) {
 	// The index where the median is supposed to be.
 	uint mid_index = end / 2;
 	printf("MID INDEX = %d\n", mid_index);
 
 	// The median is calculated depending on whether the population is even or odd.
 	if ((end + 1) % 2 == 0) {
-		double mid1 = kthSmallest(distances, 0, end, mid_index + 1);
-		double mid2 = kthSmallest(distances, 0, end, mid_index + 2);
+		float mid1 = kthSmallest(distances, 0, end, mid_index + 1);
+		float mid2 = kthSmallest(distances, 0, end, mid_index + 2);
 
 		printf("mid1 = %f, mid2 = %f\n", mid1, mid2);
-		return (double) ((mid1 + mid2) / 2);
+		return (float) ((mid1 + mid2) / 2);
 	} else {
 		return kthSmallest(distances, 0, end, mid_index);
 	}
